Remove selfsigned cert creation
User has to create his own cert if he wants to use https.
Implementing scripts etc. to do this for the user is not worth the effort.
I will also think about a different approach to this (Maybe use JWT and a shared secret over http, as there is no personal information in the http traffic. HTTPS would only be used to stop people from controlling the leds without permission)