Add Stationary Animations, resolves #3
This commit adds Stationary Animations.
First approach was to use a IStripController implementation as some kind of middleware between the Animation and the IStripController.
Then I realized I could also just use the LEDs Array and completely remove the Strip from the Animations. Now the Animations edit the given LED Array instead of setting the LEDs directly.
This also means Animations have to set all LEDs they want to "obtain" everytime update is called. Beforehand the Animations just had to update their LEDs when they needed to (Blink for example updated every few seconds).
In order to keep some of the nice effects resulting form the only update changes I added the onResume() method to the IAnimation interface. Fire uses it to "burn" over the previous Notifiation.
But Animations are by no means perfect, there are probably many visual glitches waiting for me.

Based on the Stationary Animations I will now try to implement a Notification mirroring System. A App is already in the making to power the gathering of Notifications.

The StationaryStripController is not used and wount be, it just acts as a reminder to implement some kind of similar system for the LED Array.
Every update for every Stationary Animation splices the main LED Array and I don't realy like that.